Finally finished day 15 part 2
I had blinded myself with my own tests... it turns out the seed value is
supposed to act like it was the last value you returned, not the next one for
you to return. This is obvious in retrospect, but I was so convinced it was the
"next" value that I had written my test cases like that, so of course it
passed.

The interesting thing is that with a large enough number of inputs this doesn't
actually matter if you're just counting things that rarely happen. That's why
the "wrong" approach worked for part 1, over 40000000 examples having an extra
wrong output at the beginning didn't matter. For part 2, though, the single
wrong value ended up causing the generators to go out of sync since we were
discarding the first entry differently. This effectively caused A to be one
cycle ahead of B, which actually causes problems. Although since these values
are somewhat random it was still pretty close since the probability of these
matches are equivalent.