Ok, finished day 16 part 2
Since we know that any arrangement will map to one other arrangement we can
save these in a map and skip a round of dances. This could be further improved
by noting that the first time we see a situation like this we've entered a loop
and then simply take the number of moves we have left mod the loop length, but
I didn't bother to do that since we could just burn a bit of cpu instead.