2020 day 12 part 2
I spent a lot of time misunderstanding the question and assuming the waypoint
was fixed to the world. As it turns out it is fixed relative to the ship (which
the instructions do say, but that didn't seem like a "waypoint" to me).

The vector bits from part 1 did come in handy here once I knew what the
question was *actually* asking.